requirements.txt
torch>=2.1.0
transformers>=4.36.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
datasets>=2.16.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
